Move AdBlockingRecoveryApp to its own folder.
